requirements.txt
Django==4.0.2
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
celery==5.2.3
flower==1.0.0
redis==4.1.4
environ==1.0
